ai_travel_agent/requirements.txt
chainlit
openai[agents]
python-dotenv
